LayoutTests/platform/chromium-mac/fast/canvas/set-colors-expected.txt

Test of various canvas graphics context calls for setting colors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFillStyle('transparent') is transparent
PASS testFillStyle('blue') is blue
PASS testFillStyle('#FF0000') is red
PASS testFillStyle('#f00') is red
PASS testFillStyle('rgb(255, 0, 0)') is red
PASS testFillStyle('rgba(255, 0, 0, 1)') is red
PASS testFillStyle('rgba(255, 0, 0, 0.8)') is translucentRed
PASS testFillStyle('rgba(255, 0, 0, 0)') is transparent
PASS testFillGradient('transparent') is transparent
PASS testFillGradient('blue') is blue
PASS testFillGradient('#FF0000') is red
PASS testFillGradient('#f00') is red
PASS testFillGradient('rgb(255, 0, 0)') is red
PASS testFillGradient('rgba(255, 0, 0, 1)') is red
PASS testFillGradient('rgba(255, 0, 0, 0.8)') is translucentRed
PASS testFillGradient('rgba(255, 0, 0, 0)') is transparent
PASS testSetFillColor('"blue"') is blue
PASS testSetFillColor('"#FF0000"') is red
PASS testSetFillColor('"#f00"') is red
PASS testSetFillColor('"rgb(255, 0, 0)"') is red
PASS testSetFillColor('"rgba(255, 0, 0, 1)"') is red
PASS testSetFillColor('"rgba(255, 0, 0, 0.8)"') is translucentRed
PASS testSetFillColor('"rgba(255, 0, 0, 0)"') is transparent
PASS testSetFillColor('"blue", 0.8') is translucentBlue
PASS testSetFillColor('1') is white
PASS testSetFillColor('1, 0.8') is translucentWhite
PASS testSetFillColor('0, 1, 0, 1') is green
PASS testSetFillColor('0, 1, 0, 0.8') is translucentGreen
FAIL testSetFillColor('0, 0, 0, 1, 1') should be #1a1a1a. Was #000000.
FAIL testSetFillColor('0, 0, 0, 1, 0.8') should be rgba(25, 25, 25, 0.8). Was rgba(0, 0, 0, 0.8).
PASS testSetFillColor('0, 0, 0, 1, 0') is transparent
PASS testStrokeStyle('transparent') is transparent
FAIL testStrokeStyle('blue') should be #0000ff. Was rgba(0, 0, 255, 0.9725490196078431).
FAIL testStrokeStyle('#FF0000') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeStyle('#f00') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeStyle('rgb(255, 0, 0)') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeStyle('rgba(255, 0, 0, 1)') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeStyle('rgba(255, 0, 0, 0.8)') should be rgba(255, 0, 0, 0.8). Was rgba(255, 0, 0, 0.7764705882352941).
PASS testStrokeStyle('rgba(255, 0, 0, 0)') is transparent
PASS testStrokeGradient('transparent') is transparent
FAIL testStrokeGradient('blue') should be #0000ff. Was rgba(0, 0, 255, 0.9725490196078431).
FAIL testStrokeGradient('#FF0000') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeGradient('#f00') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeGradient('rgb(255, 0, 0)') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeGradient('rgba(255, 0, 0, 1)') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testStrokeGradient('rgba(255, 0, 0, 0.8)') should be rgba(255, 0, 0, 0.8). Was rgba(255, 0, 0, 0.7764705882352941).
PASS testStrokeGradient('rgba(255, 0, 0, 0)') is transparent
FAIL testSetStrokeColor('"blue"') should be #0000ff. Was rgba(0, 0, 255, 0.9725490196078431).
FAIL testSetStrokeColor('"#FF0000"') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testSetStrokeColor('"#f00"') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testSetStrokeColor('"rgb(255, 0, 0)"') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testSetStrokeColor('"rgba(255, 0, 0, 1)"') should be #ff0000. Was rgba(255, 0, 0, 0.9725490196078431).
FAIL testSetStrokeColor('"rgba(255, 0, 0, 0.8)"') should be rgba(255, 0, 0, 0.8). Was rgba(255, 0, 0, 0.7764705882352941).
PASS testSetStrokeColor('"rgba(255, 0, 0, 0)"') is transparent
FAIL testSetStrokeColor('"blue", 0.8') should be rgba(0, 0, 255, 0.8). Was rgba(0, 0, 255, 0.7764705882352941).
FAIL testSetStrokeColor('1') should be #ffffff. Was rgba(255, 255, 255, 0.9725490196078431).
FAIL testSetStrokeColor('1, 0.8') should be rgba(255, 255, 255, 0.8). Was rgba(255, 255, 255, 0.7764705882352941).
FAIL testSetStrokeColor('0, 1, 0, 1') should be #00ff00. Was rgba(0, 255, 0, 0.9725490196078431).
FAIL testSetStrokeColor('0, 1, 0, 0.8') should be rgba(0, 255, 0, 0.8). Was rgba(0, 255, 0, 0.7764705882352941).
FAIL testSetStrokeColor('0, 0, 0, 1, 1') should be #1a1a1a. Was rgba(0, 0, 0, 0.9725490196078431).
FAIL testSetStrokeColor('0, 0, 0, 1, 0.8') should be rgba(25, 25, 25, 0.8). Was rgba(0, 0, 0, 0.7764705882352941).
PASS testSetStrokeColor('0, 0, 0, 1, 0') is transparent
PASS successfullyParsed is true

TEST COMPLETE
